Use reexports to fix ibc module dependencies on ibc-base
