+++ modules/base/src/ics04_channel/mod.rs
@@ -2,11 +2,9 @@
 //! applications and the chains those applications are built upon.
 
 pub mod channel;
+pub mod commitment;
 pub mod context;
 pub mod error;
-
 pub mod packet;
-
-pub mod commitment;
 mod version;
 pub use version::Version;
