+++ modules/src/core/ics02_client/client_def.rs
@@ -5,7 +5,7 @@ use crate::core::ics02_client::client_state::{AnyClientState, ClientState};
 use crate::core::ics02_client::client_type::ClientType;
 use crate::core::ics02_client::context::ConsensusReader;
 use crate::core::ics02_client::error::Error;
-use crate::core::ics02_client::header::{AnyHeader, Header};
+use crate::core::ics02_client::header::AnyHeader;
 use crate::core::ics03_connection::connection::ConnectionEnd;
 use crate::core::ics04_channel::channel::ChannelEnd;
 use crate::core::ics04_channel::commitment::{AcknowledgementCommitment, PacketCommitment};
@@ -23,149 +23,7 @@ use ibc_client_tendermint::client_def::TendermintClient;
 #[cfg(any(test, feature = "mocks"))]
 use crate::mock::client_def::MockClient;
 
-pub trait ClientDef: Clone {
-    type Header: Header;
-    type ClientState: ClientState;
-    type ConsensusState: ConsensusState;
-
-    fn check_header_and_update_state(
-        &self,
-        ctx: &dyn ConsensusReader,
-        client_id: ClientId,
-        client_state: Self::ClientState,
-        header: Self::Header,
-    ) -> Result<(Self::ClientState, Self::ConsensusState), Error>;
-
-    fn verify_upgrade_and_update_state(
-        &self,
-        client_state: &Self::ClientState,
-        consensus_state: &Self::ConsensusState,
-        proof_upgrade_client: MerkleProof,
-        proof_upgrade_consensus_state: MerkleProof,
-    ) -> Result<(Self::ClientState, Self::ConsensusState), Error>;
-
-    /// Verification functions as specified in:
-    /// <https://github.com/cosmos/ibc/tree/master/spec/core/ics-002-client-semantics>
-    ///
-    /// Verify a `proof` that the consensus state of a given client (at height `consensus_height`)
-    /// matches the input `consensus_state`. The parameter `counterparty_height` represent the
-    /// height of the counterparty chain that this proof assumes (i.e., the height at which this
-    /// proof was computed).
-    #[allow(clippy::too_many_arguments)]
-    fn verify_client_consensus_state(
-        &self,
-        client_state: &Self::ClientState,
-        height: Height,
-        prefix: &CommitmentPrefix,
-        proof: &CommitmentProofBytes,
-        root: &CommitmentRoot,
-        client_id: &ClientId,
-        consensus_height: Height,
-        expected_consensus_state: &dyn ConsensusState,
-    ) -> Result<(), Error>;
-
-    /// Verify a `proof` that a connection state matches that of the input `connection_end`.
-    #[allow(clippy::too_many_arguments)]
-    fn verify_connection_state(
-        &self,
-        client_state: &Self::ClientState,
-        height: Height,
-        prefix: &CommitmentPrefix,
-        proof: &CommitmentProofBytes,
-        root: &CommitmentRoot,
-        connection_id: &ConnectionId,
-        expected_connection_end: &ConnectionEnd,
-    ) -> Result<(), Error>;
-
-    /// Verify a `proof` that a channel state matches that of the input `channel_end`.
-    #[allow(clippy::too_many_arguments)]
-    fn verify_channel_state(
-        &self,
-        client_state: &Self::ClientState,
-        height: Height,
-        prefix: &CommitmentPrefix,
-        proof: &CommitmentProofBytes,
-        root: &CommitmentRoot,
-        port_id: &PortId,
-        channel_id: &ChannelId,
-        expected_channel_end: &ChannelEnd,
-    ) -> Result<(), Error>;
-
-    /// Verify the client state for this chain that it is stored on the counterparty chain.
-    #[allow(clippy::too_many_arguments)]
-    fn verify_client_full_state<U>(
-        &self,
-        client_state: &Self::ClientState,
-        height: Height,
-        prefix: &CommitmentPrefix,
-        proof: &CommitmentProofBytes,
-        root: &CommitmentRoot,
-        client_id: &ClientId,
-        expected_client_state: &dyn ClientState<UpgradeOptions = U>,
-    ) -> Result<(), Error>;
-
-    /// Verify a `proof` that a packet has been commited.
-    #[allow(clippy::too_many_arguments)]
-    fn verify_packet_data(
-        &self,
-        ctx: &dyn ChannelMetaReader,
-        client_state: &Self::ClientState,
-        height: Height,
-        connection_end: &ConnectionEnd,
-        proof: &CommitmentProofBytes,
-        root: &CommitmentRoot,
-        port_id: &PortId,
-        channel_id: &ChannelId,
-        sequence: Sequence,
-        commitment: PacketCommitment,
-    ) -> Result<(), Error>;
-
-    /// Verify a `proof` that a packet has been commited.
-    #[allow(clippy::too_many_arguments)]
-    fn verify_packet_acknowledgement(
-        &self,
-        ctx: &dyn ChannelMetaReader,
-        client_state: &Self::ClientState,
-        height: Height,
-        connection_end: &ConnectionEnd,
-        proof: &CommitmentProofBytes,
-        root: &CommitmentRoot,
-        port_id: &PortId,
-        channel_id: &ChannelId,
-        sequence: Sequence,
-        ack: AcknowledgementCommitment,
-    ) -> Result<(), Error>;
-
-    /// Verify a `proof` that of the next_seq_received.
-    #[allow(clippy::too_many_arguments)]
-    fn verify_next_sequence_recv(
-        &self,
-        ctx: &dyn ChannelMetaReader,
-        client_state: &Self::ClientState,
-        height: Height,
-        connection_end: &ConnectionEnd,
-        proof: &CommitmentProofBytes,
-        root: &CommitmentRoot,
-        port_id: &PortId,
-        channel_id: &ChannelId,
-        sequence: Sequence,
-    ) -> Result<(), Error>;
-
-    /// Verify a `proof` that a packet has not been received.
-    #[allow(clippy::too_many_arguments)]
-    fn verify_packet_receipt_absence(
-        &self,
-        ctx: &dyn ChannelMetaReader,
-        client_state: &Self::ClientState,
-        height: Height,
-        connection_end: &ConnectionEnd,
-        proof: &CommitmentProofBytes,
-        root: &CommitmentRoot,
-        port_id: &PortId,
-        channel_id: &ChannelId,
-        sequence: Sequence,
-    ) -> Result<(), Error>;
-}
+pub use ibc_base::ics02_client::client_def::*;
 
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub enum AnyClient {
