+++ modules/src/core/ics02_client/client_state.rs
@@ -1,4 +1,3 @@
-use core::marker::{Send, Sync};
 use core::time::Duration;
 
 use ibc_proto::google::protobuf::Any;
@@ -19,43 +18,10 @@ use crate::mock::client_state::MockClientState;
 use crate::prelude::*;
 use crate::Height;
 
+pub use ibc_base::ics02_client::client_state::*;
+
 pub const TENDERMINT_CLIENT_STATE_TYPE_URL: &str = "/ibc.lightclients.tendermint.v1.ClientState";
 pub const MOCK_CLIENT_STATE_TYPE_URL: &str = "/ibc.mock.ClientState";
-
-pub trait ClientState: core::fmt::Debug + Send + Sync {
-    /// Client-specific options for upgrading the client
-    type UpgradeOptions;
-
-    /// Return the chain identifier which this client is serving (i.e., the client is verifying
-    /// consensus states from this chain).
-    fn chain_id(&self) -> ChainId;
-
-    /// Type of client associated with this state (eg. Tendermint)
-    fn client_type(&self) -> ClientType;
-
-    /// Latest height of consensus state
-    fn latest_height(&self) -> Height;
-
-    /// Freeze status of the client
-    fn is_frozen(&self) -> bool {
-        self.frozen_height().is_some()
-    }
-
-    /// Frozen height of the client
-    fn frozen_height(&self) -> Option<Height>;
-
-    /// Helper function to verify the upgrade client procedure.
-    /// Resets all fields except the blockchain-specific ones,
-    /// and updates the given fields.
-    fn upgrade(
-        &mut self,
-        upgrade_height: Height,
-        upgrade_options: Self::UpgradeOptions,
-        chain_id: ChainId,
-    );
-
-    fn encode_vec(&self) -> Result<Vec<u8>, Error>;
-}
 
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
 #[serde(tag = "type")]
