+++ modules/src/core/ics02_client/context.rs
@@ -2,7 +2,7 @@
 //! that any host chain must implement to be able to process any `ClientMsg`. See
 //! "ADR 003: IBC protocol implementation" for more details.
 
-use crate::core::ics02_client::client_consensus::{AnyConsensusState, ConsensusState};
+use crate::core::ics02_client::client_consensus::AnyConsensusState;
 use crate::core::ics02_client::client_state::AnyClientState;
 use crate::core::ics02_client::client_type::ClientType;
 use crate::core::ics02_client::error::{Error, ErrorDetail};
@@ -10,7 +10,8 @@ use crate::core::ics02_client::handler::ClientResult::{self, Create, Update, Upg
 use crate::core::ics24_host::identifier::ClientId;
 use crate::timestamp::Timestamp;
 use crate::Height;
-use std::prelude::v1::Box;
+
+pub use ibc_base::ics02_client::context::*;
 
 /// Defines the read-only part of ICS2 (client functions) context.
 pub trait ClientReader {
@@ -79,78 +80,38 @@ pub trait ClientReader {
     fn client_counter(&self) -> Result<u64, Error>;
 }
 
-pub trait ConsensusReader {
-    fn consensus_state(
-        &self,
-        client_id: &ClientId,
-        height: Height,
-    ) -> Result<Box<dyn ConsensusState>, Error>;
-
-    /// Similar to `consensus_state`, attempt to retrieve the consensus state,
-    /// but return `None` if no state exists at the given height.
-    fn maybe_consensus_state(
-        &self,
-        client_id: &ClientId,
-        height: Height,
-    ) -> Result<Option<Box<dyn ConsensusState>>, Error> {
-        match self.consensus_state(client_id, height) {
-            Ok(cs) => Ok(Some(cs)),
-            Err(e) => match e.detail() {
-                ErrorDetail::ConsensusStateNotFound(_) => Ok(None),
-                _ => Err(e),
-            },
-        }
-    }
-
-    /// Search for the lowest consensus state higher than `height`.
-    fn next_consensus_state(
-        &self,
-        client_id: &ClientId,
-        height: Height,
-    ) -> Result<Option<Box<dyn ConsensusState>>, Error>;
-
-    /// Search for the highest consensus state lower than `height`.
-    fn prev_consensus_state(
-        &self,
-        client_id: &ClientId,
-        height: Height,
-    ) -> Result<Option<Box<dyn ConsensusState>>, Error>;
-
-    /// Returns the current timestamp of the local chain.
-    fn host_timestamp(&self) -> Timestamp;
-}
-
-impl<T: ClientReader> ConsensusReader for T {
-    fn consensus_state(
-        &self,
-        client_id: &ClientId,
-        height: Height,
-    ) -> Result<Box<dyn ConsensusState>, Error> {
-        ClientReader::consensus_state(self, client_id, height).map(|cs| cs.boxed_dyn())
-    }
-
-    fn next_consensus_state(
-        &self,
-        client_id: &ClientId,
-        height: Height,
-    ) -> Result<Option<Box<dyn ConsensusState>>, Error> {
-        ClientReader::next_consensus_state(self, client_id, height)
-            .map(|cs| cs.map(AnyConsensusState::boxed_dyn))
-    }
-
-    fn prev_consensus_state(
-        &self,
-        client_id: &ClientId,
-        height: Height,
-    ) -> Result<Option<Box<dyn ConsensusState>>, Error> {
-        ClientReader::prev_consensus_state(self, client_id, height)
-            .map(|cs| cs.map(AnyConsensusState::boxed_dyn))
-    }
-
-    fn host_timestamp(&self) -> Timestamp {
-        ClientReader::host_timestamp(self)
-    }
-}
+// FIXME(hu55a1n1)
+// impl<T: ClientReader> ConsensusReader for T {
+//     fn consensus_state(
+//         &self,
+//         client_id: &ClientId,
+//         height: Height,
+//     ) -> Result<Box<dyn ConsensusState>, Error> {
+//         ClientReader::consensus_state(self, client_id, height).map(|cs| cs.boxed_dyn())
+//     }
+//
+//     fn next_consensus_state(
+//         &self,
+//         client_id: &ClientId,
+//         height: Height,
+//     ) -> Result<Option<Box<dyn ConsensusState>>, Error> {
+//         ClientReader::next_consensus_state(self, client_id, height)
+//             .map(|cs| cs.map(AnyConsensusState::boxed_dyn))
+//     }
+//
+//     fn prev_consensus_state(
+//         &self,
+//         client_id: &ClientId,
+//         height: Height,
+//     ) -> Result<Option<Box<dyn ConsensusState>>, Error> {
+//         ClientReader::prev_consensus_state(self, client_id, height)
+//             .map(|cs| cs.map(AnyConsensusState::boxed_dyn))
+//     }
+//
+//     fn host_timestamp(&self) -> Timestamp {
+//         ClientReader::host_timestamp(self)
+//     }
+// }
 
 /// Defines the write-only part of ICS2 (client functions) context.
 pub trait ClientKeeper {
