+++ modules/src/core/ics02_client/mod.rs
@@ -3,13 +3,14 @@
 pub mod client_consensus;
 pub mod client_def;
 pub mod client_state;
-pub mod client_type;
 pub mod context;
-pub mod error;
 pub mod events;
 pub mod handler;
 pub mod header;
-pub mod height;
 pub mod misbehaviour;
 pub mod msgs;
-pub mod trust_threshold;
+
+pub use ibc_base::ics02_client::client_type;
+pub use ibc_base::ics02_client::error;
+pub use ibc_base::ics02_client::height;
+pub use ibc_base::ics02_client::trust_threshold;
