+++ modules/src/core/ics03_connection/mod.rs
@@ -1,12 +1,13 @@
 //! ICS 03: Connection implementation for connecting a client
 //! on the local chain with a client on a remote chain.
 
-pub mod connection;
 /// Context definitions (dependencies for the protocol).
 pub mod context;
-pub mod error;
 pub mod events;
 /// Message processing logic (protocol) for ICS 03.
 pub mod handler;
 pub mod msgs;
-pub mod version;
+
+pub use ibc_base::ics03_connection::connection;
+pub use ibc_base::ics03_connection::error;
+pub use ibc_base::ics03_connection::version;
