--- modules/src/core/ics23_commitment/mock.rs
@@ -1,1 +1,0 @@
-
