--- modules/src/core/ics23_commitment/mod.rs
@@ -1,8 +1,0 @@
-//! ICS 23: Commitment implementation of a cryptographic scheme that verifies
-//! state transitions between chains.
-
-pub mod commitment;
-pub mod error;
-pub mod merkle;
-pub mod mock;
-pub mod specs;
