--- modules/src/core/ics24_host/mod.rs
@@ -1,9 +1,0 @@
-//! ICS 24: Host defines the minimal set of interfaces that a
-//! state machine hosting an IBC-enabled chain must implement.
-
-pub use path::{ClientUpgradePath, Path, IBC_QUERY_PATH, SDK_UPGRADE_QUERY_PATH};
-
-pub mod error;
-pub mod identifier;
-pub mod path;
-pub mod validate;
