+++ modules/src/core/mod.rs
@@ -5,6 +5,7 @@ pub mod ics02_client;
 pub mod ics03_connection;
 pub mod ics04_channel;
 pub mod ics05_port;
-pub mod ics23_commitment;
-pub mod ics24_host;
 pub mod ics26_routing;
+
+pub use ibc_base::ics23_commitment;
+pub use ibc_base::ics24_host;
